Se agrega consulta de reviews y calculo de ranking por libro

+++ src/main/java/com/unir/books/data/model/Book.java
@@ -60,7 +60,7 @@ public class Book {
     private Date datePublished;
 
     @Column(name ="ranking")
-    private Integer ranking;
+    private Double ranking;
 
     @Column(name ="type")
     private String type;
